Don't create csgs until they're needed.
