+++ Elements/src/Geometry/Solids/Extrude.cs
@@ -88,7 +88,6 @@ namespace Elements.Geometry.Solids
         private void UpdateGeometry()
         {
             this._solid = Kernel.Instance.CreateExtrude(this._profile, this._height, this._direction);
-            this._csg = this._solid.ToCsg();
         }
     }
 }
